fastapi_crud/requirements.txt
fastapi
uvicorn

sqlalchemy
psycopg2-binary
pydantic

python-dotenv

httpx
pytest
pytest-cov

prometheus-fastapi-instrumentator

